+Simple test of camera preview.
In normal case it would be going through multiple passes instead of just one.
Lightning is messed up because of transform buffers in objects, we would need to do a lot of work to override all transform buffers before drawing, and that would start problems. For now we just leave it like this and come back with better setup